Create layout component, restructure components
