+++ src/components/footer/Footer.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+export const Footer = () => {
+  return (
+    <footer className="container-fluid py-5 bg-dark text-white">
+      <div className="container-fluid text-center">
+        <small>
+          Copyright &copy; Designed with <i className="far fa-heart" /> by
+          PyLadies Bratislava
+        </small>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
