+++ src/components/nav-link/index.jsx
@@ -1,0 +1,1 @@
+export * from "./NavLink";
