+++ src/components/page-layout/index.jsx
@@ -1,0 +1,1 @@
+export * from "./PageLayout";
