+++ src/pages/index.jsx
@@ -1,3 +1,4 @@
 export * from "./About";
 export * from "./MeetUps";
 export * from "./Home";
+export * from "./Events";
